refactor gallery page into components and hook
